Inline lwAFTR helper ICMP library into lwaftr.lua
